% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GbusterMean.R
\name{GbusterMean}
\alias{GbusterMean}
\title{Finds the parameters of a gamma distribution that best fit a given expected mean and range}
\usage{
GbusterMean(mean, range, pct = 0.975, start)
}
\arguments{
\item{mean}{The expected value}

\item{range}{The expected range containing the mean}

\item{pct}{One's confidence in the above range, e.g., .95}

\item{start}{Starting values for the shape and scale parameters of a gamma function}
}
\value{
A list of output.  The first list item $par contains the parameters describing the shape
 and scale of a gamma distribution.  The remaining list items describe the convergence of
 the search algorithm.
}
\description{
Finds the parameters of a gamma distribution that best fit a given expected mean and range
}
\examples{
GbusterMean(20, 30)
GbusterMean(mean=200, range=400, pct=.80, start=c(10,.05))
}
