% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GbusterMinMax.R
\name{GbusterMinMax}
\alias{GbusterMinMax}
\title{Find the parameters of a gamma distribution that best fit a given expected minumum and maximum}
\usage{
GbusterMinMax(lower, upper, pct = 0.95, start)
}
\arguments{
\item{lower}{An expected lower bound}

\item{upper}{An expected upper bound}

\item{pct}{One's confidence in the above bounds, e.g., .95}

\item{start}{Starting values for the shape and scale parameters of a gamma function}
}
\value{
A list of output.  The first list item $par contains the parameters describing the shape
 and scale of a gamma distribution.  The remaining items describe the convergence of
 the search algorithm.
}
\description{
Find the parameters of a gamma distribution that best fit a given expected minumum and maximum
}
\examples{
GbusterMinMax(20, 30)
GbusterMinMax(400, 1500, pct=.80, start=c(200,2))

# Note that this function won't operate on vectors of inputs.  So, if you're trying to operate on a
# data frame such as the following...
min <- c(20, 30, 40)
max <- c(40, 35, 80)
df <- data.frame(min, max)
# # ...then you'll need to apply this function row-by-row such as in the following loop (or via a creative
# application of something from the apply() family):
for(i in 1:nrow(df)) {
gamma.parameters <- GbusterMinMax(df$min[i], df$max[i])  # Find the parameters of a gamma distribution for agiven min and max
df$Freq[i] <- gamma.parameters$par[1] / gamma.parameters$par[2]
df$Weight[i] <- gamma.parameters$par[2]
}
}
